add random altair fork tests for better translation coverage + pr review
