+++ tests/core/pyspec/eth2spec/test/helpers/rewards.py
@@ -3,11 +3,15 @@ from lru import LRU
 
 from eth2spec.phase0 import spec as spec_phase0
 from eth2spec.test.context import is_post_altair
-from eth2spec.test.helpers.attestations import cached_prepare_state_with_attestations
 from eth2spec.test.helpers.state import (
     next_epoch,
+)
+from eth2spec.test.helpers.random import (
     set_some_new_deposits, exit_random_validators, slash_random_validators,
     randomize_state,
+)
+from eth2spec.test.helpers.attestations import (
+    cached_prepare_state_with_attestations,
 )
 from eth2spec.utils.ssz.ssz_typing import Container, uint64, List
